+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "devstart": "nodemon server.js",
-    "start":"node server.js"
+    "start": "node server.js"
   },
   "author": "Rob Miles",
   "license": "ISC",
@@ -22,6 +22,7 @@
     "mqtt": "^4.2.8",
     "passport": "^0.4.1",
     "passport-local": "^1.0.0",
+    "qrcode": "^1.5.1",
     "slugify": "^1.4.6"
   },
   "devDependencies": {
