+++ routes/GUIDcommand.js
