+++ views/device.ejs
@@ -63,6 +63,13 @@
             <a href="/pixel" class="btn btn-success mb-4">Pixels</a>
         </div>
 
+        <h2 class="mb-4 mt-4">
+            Create Controls
+        </h2>
+        <div class="container mb-4 mt-4">
+            <a href="/pixel/control/<%= device.name %>" class="btn btn-success mb-4">Pixels</a>
+        </div>
+
         <form action="/devices/sendconsolecommand/<%= device.name %>"" method="POST">
             <div class="form-group mt-4">
                 <label for="consoleCommand">Console Command</label>
